refactor(engineer_loop): split tests_verification.rs again (#1266)
tests_verification.rs 759 -> 395 LOC; new tests_verification_more.rs (~383).
